requirements.txt
crewai
langchain
langchain-community
python-dotenv
pydantic
autogen
openai
streamlit
langchain_groq
groq